+++ src/components/control/TimeFrame.story.vue
@@ -1,22 +1,22 @@
 <template>
-    <Story title="Form/Control/TimeFrameField" :layout="{ type: 'grid', width: '800px' }">
+    <Story title="Form/Control/TimeFrame" :layout="{ type: 'grid', width: '800px' }">
         <Variant title="Default" :init-state="stateDefault" v-slot="{ state }">
             <div class="max-w-[400px] mx-auto p-2">
-                <TimeFrameField v-bind="state" v-model="state.value" />
+                <TimeFrame v-bind="state" v-model="state.value" />
             </div>
         </Variant>
 
         <Variant title="Between" :init-state="stateBetween" v-slot="{ state }">
             <div class="max-w-[400px] mx-auto p-2">
-                <TimeFrameField v-bind="state" v-model="state.value" />
+                <TimeFrame v-bind="state" v-model="state.value" />
             </div>
         </Variant>
         
         <Variant title="Sizes" :init-state="stateSizes" v-slot="{ state }">
             <div class="max-w-[400px] mx-auto flex flex-col gap-4 p-2">
-                <TimeFrameField v-bind="state" size="sm" placeholder="32px height" v-model="state.valueSM" />
-                <TimeFrameField v-bind="state" placeholder="40px height" v-model="state.valueMD" />
-                <TimeFrameField v-bind="state" size="lg" placeholder="56px height" v-model="state.valueLG" />
+                <TimeFrame v-bind="state" size="sm" placeholder="32px height" v-model="state.valueSM" />
+                <TimeFrame v-bind="state" placeholder="40px height" v-model="state.valueMD" />
+                <TimeFrame v-bind="state" size="lg" placeholder="56px height" v-model="state.valueLG" />
             </div>
         </Variant>
     </Story>
@@ -24,7 +24,7 @@
 
 <script lang="ts" setup>
 import { ref } from 'vue';
-import TimeFrameField from './TimeFrameField.vue';
+import TimeFrame from './TimeFrame.vue';
 
 function stateDefault() {
     return {