+++ CHANGELOG.md
@@ -15,6 +15,7 @@ miru.ink / Changelog
 - Add: New `escape` property on `DialogStd` component to allow closing the modal using the ESC key.
 - Add: New wobble animation when trying to close a static modal by an click outside.
 - Add: `CheckCircle` / `XCircle` lucide icon components.
+- Add: `TimeSlots` input control field component, supporting booking a timeslot.
 - Update: Change the `AvatarGroup` limit class and label when not enough space is available.
 - Update: `package.json` dependencies.
 - Update: Decrease height (by lower the padding) on `BadgeStd` component.
@@ -26,6 +27,8 @@ miru.ink / Changelog
 - Update: Apply new 32 / 40 / 56 px height system on `NumberField` control field.
 - Update: Apply new 32 / 40 / 56 px height system on `PasswordField` control field.
 - Update: Apply new 32 / 40 / 56 px height system on `SelectField` control field.
+- Update: Rename `TimeFrameField` to `TimeFrame`.
+- Update: Add `type` option to `now()` utility function to change returning value.
 - Fix: Add media-query based `max-width` stylings on `DialogStd` to support some space to screen-borders.
 - Fix: Demo-Icons on histoire stories.
 - Fix: Minor styling changes on `TooltipStd` component.